chore: move example test in example dir
