+++ example/ringbuffer.yaml
@@ -1,0 +1,5 @@
+RingBufferCfg:
+  name: ringbuffer
+  type: RingBufferFloat
+  params: 
+    size: 512
